requirements.txt
numpy
astropy
click
sdss-marvin